Update entrypoint to check for existing WP install.

+++ dev-entrypoint.sh
@@ -15,11 +15,21 @@ chmod +x /usr/local/bin/cron-env.sh
 # Reload the config.
 kill -USR2 1
 
-# Run WP-CLI commands now that PHP is configured.
-wp --allow-root --quiet option set blog_public 0
-wp --allow-root --quiet plugin install query-monitor --activate --force
-wp --allow-root --quiet config shuffle-salts
-wp --allow-root --quiet rewrite flush
+# Finalize setup.
+if [ -f /var/www/html/wp-config.php ]; then
+  # There's an existing WP install.
+  # Run WP-CLI commands.
+  wp --allow-root --quiet option set blog_public 0
+  wp --allow-root --quiet plugin install query-monitor --activate --force
+  wp --allow-root --quiet config shuffle-salts
+  wp --allow-root --quiet rewrite flush
+  # Exec php-fpm as we won't run the default entry point.
+  exec php-fpm
+else
+  # No existing WP install.
+  # Run the base images entrypoint.
+  exec /usr/local/bin/docker-entrypoint.sh "php-fpm"
+fi
 
-# Run the base  images entrypoint.
-exec /usr/local/bin/docker-entrypoint.sh "php-fpm"
+
+
